feat(upload): move processing to a dedicated dir
